+++ src/index.ts
@@ -6,4 +6,5 @@ export * from "./data/local-ip-ranges.js";
 
 export * from "./functions/escape-sql-like-string.js";
 export * from "./functions/is-local-ip-address.js";
+export * from "./functions/pad-string-with-null.js";
 export * from "./functions/split-string-into-chunks.js";