+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Added `splitStringIntoChunks` function.
 * Added `localIpRanges` array.
 * Added `isLocalIpAddress` function.
+* Added `padStringWithNull` function.
 
 ## 1.0.0
 Initial release.